db/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
qdrant-client==1.7.0
sentence-transformers==2.2.2
anthropic==0.7.8
pydantic==2.5.0
python-multipart==0.0.6
asyncpg==0.29.0
asyncio-mqtt==0.16.1
python-jose==3.3.0
passlib==1.7.4
httpx==0.25.2
